+++ broker/broker_test.go
@@ -33,11 +33,12 @@ type brokerTst struct {
 }
 
 type evt struct {
-	t   events.Type
-	ctx context.Context
-	sid uint64
-	id  string
-	cid string
+	t      events.Type
+	ctx    context.Context
+	sid    uint64
+	id     string
+	cid    string
+	txHash string
 }
 
 func getBroker(t *testing.T) *brokerTst {
@@ -59,10 +60,11 @@ func (b brokerTst) randomEvt() *evt {
 		idString = ctxV
 	}
 	return &evt{
-		t:   events.All,
-		ctx: b.ctx,
-		id:  idString,
-		cid: "testchain",
+		t:      events.All,
+		ctx:    b.ctx,
+		id:     idString,
+		cid:    "testchain",
+		txHash: "testTxHash",
 	}
 }
 
@@ -721,6 +723,10 @@ func (e evt) TraceID() string {
 
 func (e evt) ChainID() string {
 	return e.cid
+}
+
+func (e evt) TxHash() string {
+	return e.txHash
 }
 
 func (e evt) StreamMessage() *eventspb.BusEvent {