+++ processor/tx.go
@@ -5,9 +5,10 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/tendermint/tendermint/crypto/tmhash"
+
 	"code.vegaprotocol.io/protos/commands"
 	commandspb "code.vegaprotocol.io/protos/vega/commands/v1"
-	"code.vegaprotocol.io/vega/libs/crypto"
 	"code.vegaprotocol.io/vega/txn"
 	wcrypto "code.vegaprotocol.io/vegawallet/crypto"
 
@@ -327,7 +328,7 @@ func (t Tx) Party() string {
 }
 
 func (t Tx) Hash() []byte {
-	return crypto.Hash(t.originalTx)
+	return tmhash.Sum(t.originalTx)
 }
 
 func (t Tx) Signature() []byte {